requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
python-telegram-bot==20.7
python-dotenv==1.0.0